로그인 기능과..그에 따른 상호작용들 구현....날씨도 조금 수정함...

+++ components/Delbtn.tsx
@@ -6,26 +6,29 @@ import { useRouter } from 'next/navigation';
 export default function Delbtn(props:any) {
   let router = useRouter()
 
-  // const requestConfig: AxiosRequestConfig = {
-  //   data: {
-  //     _id: props._id
-  //   },
-
-  // };
-
   return(
     <div>
      <button onClick={(e)=>{
-      e.preventDefault()
-      // axios.delete('/api/delete',requestConfig)
-      fetch('/api/delete',{
-        method:'DELETE',
-        body: props._id
-      }).then((result)=>{
-        router.push('/')
-      }).then(()=>{
-        router.refresh()
-      })
+      if(props.curuser == props.result.writer) {
+        e.preventDefault()
+
+        fetch('/api/delete',{
+          method:'DELETE',
+          body: props._id
+        }).then((result)=>{
+          router.push('/')
+        }).then(()=>{
+          router.refresh()
+        })
+      } else {
+        e.preventDefault()
+        alert("글 수정, 삭제는 본인만이 가능합니다.")
+      }
+
+
+
+
+
      }} className={styles.modifyp}>삭제 ⌫</button>
     </div>
   )
